\name{addSpecial}
\alias{remSpecial}
\alias{addSpecial}

\title{Add or remove a "Special" menue to the RGui}
\description{
  The function is used to add or remove a user menue "Special" to the RGui.
}
\usage{
remSpecial()
addSpecial()
}

\details{
  This function works on Windows only.
}
\value{
  No value is returned
}

\seealso{\code{\link{sethelp}}}

\keyword{misc}
