% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammStd-function.R
\name{ammStd}
\alias{ammStd}
\title{Ammonia standard caclulation for Washington, DC}
\usage{
ammStd(month, temperature, pH)
}
\arguments{
\item{month, }{temperaure, pH}
}
\description{
This function calculates the ammonia standards based on season, pH, and water temperature.
}
\examples{
cat_function()
}
\keyword{Ammonia}
\keyword{wqs}
