% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_season.R
\name{assign_season}
\alias{assign_season}
\title{Assign seasons to sampling data}
\usage{
assign_season(data, season = NULL, season_names = NULL,
  season_start = NULL, abb = FALSE)
}
\arguments{
\item{data}{a vector of POSIXct dates}

\item{season}{A list of seasons. Months (1-12) are assigned to different groups based on user preference. Defaults to 12 months, starting with January. Must assign a minimum of two seasons}

\item{season_names}{A string vector of season names. The number of season names must match the length of the \code{season} list. A minimum of two seasons must be assigned (e.g., 'Wet', 'Dry'). Defaults to 12 months, starting with January. The number of season names must match the number of seasons}

\item{season_start}{defaults to 12 months, starting with January}

\item{abb}{logical, should abbreviations for month names be used? Defaults to \code{FALSE}}
}
\value{
Returns a vector of ordered season factors.
}
\description{
Assign seasons to sampling data on a monthly basis or user-defined basis
}
\details{
A function used to assign seasons to sampling data and to order the seasons. This function also allows the user to designate which season should be first.
}
\author{
Julie Padilla
}
\concept{wrangling}
