% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthType-function.R
\name{depthType}
\alias{depthType}
\title{Water Column Depth Designator (surface or subsurface) for samples}
\usage{
depthType(startDepth, endDepth, cutoff)
}
\arguments{
\item{startDepth}{column containing starting depth of sampling feature}

\item{endDepth}{column containing ending depth of sampling feature}

\item{cutoff}{numeric value to specificy where surface/subsurface cutoff should occur.}
}
\description{
This function returns a vector designating whether a sample is a surface or subsurface sample depending a single cutoff depth.
}
\examples{
depthType()
}
\keyword{column,}
\keyword{subsurface}
\keyword{surface,}
\keyword{water}
