% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mulitple_dates.R
\name{filter_multiple_dates}
\alias{filter_multiple_dates}
\title{Filter a data.frame for multiple dates}
\usage{
filter_multiple_dates(x, date_var, date_list, use_names = T)
}
\arguments{
\item{x}{a data.frame}

\item{date_var}{chr string of the name for the datetime column}

\item{date_list}{a list of dates in the format of c(start, end). Dates will be used to filter \code{data.frame}, inclusive. Dates must be in \code{POSIXct} format.}

\item{use_names}{logical If \code{TRUE} a new column will be added to label events based on names assigned to elements in \code{date_list}}
}
\value{
Returns a data.frame
}
\description{
Filter a data.frame for multiple time periods
}
\details{
This function allows the user to filter a single data.frame for multiple time periods.
}
\examples{
\dontrun{
df <- data.frame(datetime = seq(from = as.POSIXct('2000-01-01')
, to = as.POSIXct('2000-02-01'), by = 'hour')
, value = rnorm(745))

events <- lapply(list(c('2000-01-02 0:00', '2000-01-02 05:00')
, c('2000-01-04 0:00', '2000-01-04 05:00')), as.POSIXct)

filter_multiple_dates(df, date_var = 'datetime', date_list = events, use_names = F)
}

}
\author{
Julie Padilla
}
