\name{limnotools-package}
\alias{limnotools-package}
\alias{limnotools}
\docType{package}
\title{
Useful functions from Workgroup Limnology in Dresden, Germany
}
\description{
This package is a growing collection of tools and functions which may be useful
especially for R beginners and in also for limnologists.
}
\details{

This package is at a very early stage of development. The documentation 
is still incomplete and function names may change as well as the name of the
package itself. Bugs are likely (e.g. in loadcalc), so please check your results.

If you are uncomfortable with this, you are invited to contribute.


\tabular{ll}{
Package: \tab limnotools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-01-22\cr
License: \tab GNU Public License 2.0 or above\cr
}

}
\author{
Thomas Petzoldt
}
\references{
\url{http://tu-dresden.de/Members/thomas.petzoldt}
}


\keyword{ package }