% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_classifier.R
\name{nd_classifier}
\alias{nd_classifier}
\title{Assign detect/non-detect to sampling data}
\usage{
nd_classifier(x)
}
\arguments{
\item{x}{a character vector values}
}
\value{
Returns a character vector
}
\description{
Assign seasons to sampling data on a monthly basis or user-defined basis
}
\details{
This function assign D/ND based on the presence of a "<" flag attached to the sampling result value.
Currently, this function is used in conjunction with \code{nd_remover}
}
\seealso{
\code{nd_remover}
}
