% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_remover.R
\name{nd_remover}
\alias{nd_remover}
\title{Remove non-detect ("<") flags from sampling data}
\usage{
nd_remover(x)
}
\arguments{
\item{x}{a character vector values}
}
\value{
Returns a numeric vector
}
\description{
Assign seasons to sampling data on a monthly basis or user-defined basis
}
\details{
This function removes the "<" flag from the front a sampling result value.
Currently, this function is used in conjunction with \code{nd_classifier}
}
\seealso{
\code{nd_classifier}
}
