% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndSub-function.R
\name{ndSub}
\alias{ndSub}
\title{A non-detect substitution function}
\usage{
ndSub(df, x, y, method)
}
\arguments{
\item{df}{data.frame or tibble containing x and y}

\item{x}{colname of the vector containing results}

\item{y}{colname of the vector containing method DLs}

\item{method}{method for handling substitutions, currently "halfDL" is the only method supported.}
}
\description{
This function allows you to pass a tibble, specify a results vector, and a detection limit vector.
Any result value below it's associated detection limit will be updated to half the detection limit value.
The result will be the original tibble with the updated results vector.
}
\examples{
ndSub()
}
\keyword{dl,}
\keyword{halfDL}
\keyword{non-detect,}
