% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normVarNames.R
\name{normVarNames}
\alias{normVarNames}
\title{Normalize Variable Names}
\usage{
normVarNames(vars, sep = "_")
}
\arguments{
\item{vars}{chr vector}

\item{sep}{a character string to separate the terms}
}
\value{
Returns a character vector
}
\description{
Normalize variable names
}
\details{
A function used to normalize unusual text formats.
}
\references{
Williams, G. J. (2011), Data Mining with Rattle and R: The Art of Excavating Data for Knowledge Discovery, Use R!, Springer.
}
\author{
Graham J. Williams
}
\concept{wrangling}
