% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2b.R
\name{r2b}
\alias{r2b}
\title{computing a norm value for the segment from point k1 to k2-1}
\usage{
r2b(k1, k2, x, y)
}
\arguments{
\item{k1}{[INTEGER] start point}

\item{k2}{[INTEGER] end point+1}

\item{x}{[REAL(?)] input x-axis array (predictor)}

\item{y}{[REAL(?)] input y-axis array (response)}
}
\value{
err [REAL] error norm calculated as the maximum pointwise geometrical distance measured perpendicular to the regression line.
}
\description{
Finds the least-squares regression line through the data points with index >= k1 and < k2.
Computes the perpendicular distance of each point from the regression line and returns the maximum.
}
\examples{
ni <- c( 1, 201, 402 )
i <- 1
k1 <- ni[i]
k2 <- ni[i+1]


r2b(k1, k2, y=t11$temper, x=t11$depth)
}

