% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_winmodel_sheet.R
\name{read_winmodel_sheet}
\alias{read_winmodel_sheet}
\title{Import WinModel Model Results From a xlsb Sheet}
\usage{
read_winmodel_sheet(sheet, wb, data.top.cell = "A10",
  header.top.cell = "B4", key = "rchres")
}
\arguments{
\item{sheet}{name of the sheet to load data from}

\item{wb}{path and workbook name of the WinModel-generated xlsb file}

\item{data.top.cell}{top left cell location where the WinModel model results start (in terms of "letter-number")}

\item{header.top.cell}{top left cell location where the WinModel header information starts (in terms of "letter-number")}

\item{key}{char, used to determine the name of the key column used by \code{tidyr::gather}}
}
\value{
Returns a \code{data.frame}
}
\description{
Import WinModel model results into R from a single worksheet in an xlsb workbook
}
\author{
Julie Padilla
}
