% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_comma.R
\name{remove_comma}
\alias{remove_comma}
\title{Remove commas in numeric values}
\usage{
remove_comma(x)
}
\arguments{
\item{x}{vector of numbers that has been imported as character because of internal commas (e.g., "1,000")}
}
\value{
Returns a numeric vector
}
\description{
Remove internal commas from numbers and convert to numeric
}
\details{
This function handles vectors of numbers that have been imported as \code{character} or \code{factor} values due to the inclusion of an internal comma (e.g. "14,300" instead of "14300"). The function removes all internal commas and then converts it to numeric.
}
\examples{
x <- c('14,300', 1, 2, 5, 7)
y <- as.factor(x)

remove_comma(x)
remove_comma(y)

}
\author{
Julie Padilla
}
