% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_serial_time_conversion.R
\name{xl_serial_time_conversion}
\alias{xl_serial_time_conversion}
\title{Excel Serial Time Conversion}
\usage{
xl_serial_time_conversion(x, unit = NA)
}
\arguments{
\item{x}{num serial date}

\item{unit}{a character string specifying a time unit or a multiple of a unit to be rounded to. Valid base units are second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year. Arbitrary unique English abbreviations as in the \code{\link[lubridate]{period()}} constructor are allowed. Rounding to multiple of units (except weeks) is supported. If no unit is specified, unit defaults to "minute".}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Create an empty US base map in ggplot
}
\details{
Converts serial dates from excel into POSIXct dates
}
\examples{
\dontrun{
xl_serial_time_conversion(40544.86)
}

}
\author{
Julie Padilla
}
\concept{
Conversion
}
