% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_base_map.R
\name{us_base_map}
\alias{us_base_map}
\title{US Base Map}
\usage{
us_base_map(incl = c("contig", "AK", "HI", "PR"), agg_county = T)
}
\arguments{
\item{incl}{Str vector to include AK, HI , and PR}

\item{agg_county}{logical, should counties be aggregated tot he state-level? Defaults to \code{TRUE}}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Create an empty US base map in ggplot
}
\details{
Creates a base map of the US with options for including AK, HI, and PR.
This function was developed from a blog post by Bob Rudis (https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/)
}
\examples{
\dontrun{
us_base_map()
}

}
\seealso{
\code{nd_remover}
}
\author{
Bob Rudis, Julie Padilla
}
\concept{
mapping
}
