% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_serial_time_conversion.R
\name{xl_serial_time_conversion}
\alias{xl_serial_time_conversion}
\title{Excel Serial Time Conversion}
\usage{
xl_serial_time_conversion(x, unit = NA)
}
\arguments{
\item{x}{num serial date}

\item{unit}{a character string specifying a time unit or a multiple of a unit to be rounded to. Valid base units are second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year. Arbitrary unique English abbreviations as in the \code{\link[lubridate]{period}} constructor are allowed. Rounding to multiple of units (except weeks) is supported. If no unit is specified, unit defaults to "minute".}
}
\value{
Returns a \code{POSIXct} object
}
\description{
Convert serial times from MS Excel to POSIXct format
}
\details{
Converts serial dates from excel into POSIXct dates. \code{\link[lubridate]{round_date}} takes a date-time object and rounds it to the nearest value of the specified time unit. For rounding date-times which is exactly halfway between two consecutive units, the convention is to round up. Note that this is in line with the behavior of R's \code{\link[base]{round.POSIXt}} function but does not follow the convention of the base \code{\link[base]{round}} function which "rounds to the even digit" per IEC 60559.
}
\examples{
\dontrun{
xl_serial_time_conversion(40544.86)
}

}
\seealso{
\code{\link[lubridate]{round_date}}
}
\author{
Julie Padilla
}
\concept{conversion}
